Upgrade to Astro v6
