+++ package.json
@@ -12,10 +12,10 @@
   },
   "dependencies": {
     "@alpinejs/persist": "^3.15.8",
-    "@astrojs/alpinejs": "^0.4.9",
-    "@astrojs/check": "^0.9.8",
-    "@astrojs/mdx": "^4.3.14",
-    "@astrojs/react": "^4.4.2",
+    "@astrojs/alpinejs": "^0.5.0",
+    "@astrojs/check": "^0.9.9",
+    "@astrojs/mdx": "^5.0.6",
+    "@astrojs/react": "^5.0.5",
     "@astrojs/rss": "^4.0.18",
     "@astrojs/sitemap": "^3.7.2",
     "@astrojs/tailwind": "^6.0.2",
@@ -31,7 +31,7 @@
     "@types/react": "^18.3.28",
     "@types/react-dom": "^18.3.7",
     "alpinejs": "^3.15.8",
-    "astro": "^5.18.1",
+    "astro": "^6.3.3",
     "astro-expressive-code": "^0.38.3",
     "astro-icon": "^1.1.5",
     "astro-seo": "^0.8.4",
