+++ src/lib/utils.ts
@@ -1,19 +1,19 @@
 import { marked } from "marked";
 
-export const conceptPagePath = (slug: string): string => {
-  return pagePath("concepts", slug);
+export const conceptPagePath = (id: string): string => {
+  return pagePath("concepts", id);
 };
 
 export const md = async (md: string): Promise<string> => {
   return marked.parseInline(md);
 };
 
-export const startPagePath = (slug: string): string => {
-  return pagePath("start", slug.substring(1));
+export const startPagePath = (id: string): string => {
+  return pagePath("start", id.substring(1));
 };
 
-const pagePath = (collection: string, slug: string): string => {
-  return `/${collection}/${slug}`;
+const pagePath = (collection: string, id: string): string => {
+  return `/${collection}/${id}`;
 };
 
 export const formatDate = (date: Date): string => {
