+++ src/pages/about.astro
@@ -1,18 +1,18 @@
 ---
-import { getEntry } from "astro:content";
-import Layout from "../layouts/Layout.astro";
+import { getEntry, render } from "astro:content";
+import FeedbackBar from "../components/FeedbackBar.astro";
 import Hero from "../components/Hero.astro";
 import HorizontalContainer from "../components/HorizontalContainer.astro";
-import Separator from "../components/Separator.astro";
-import FeedbackBar from "../components/FeedbackBar.astro";
 import Brief from "../components/mdx/Brief.astro";
+import Separator from "../components/Separator.astro";
+import Layout from "../layouts/Layout.astro";
 
 const page = await getEntry("plain", "about")!;
 
 const {
   data: { title, description },
 } = page;
-const { Content } = await page.render();
+const { Content } = await render(page);
 ---
 
 <Layout {title} {description}>
